# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-tenant SaaS API",
    "description": "API for the Multi-tenant SaaS application (DB schema-aligned)",
    "version": "0.0.1"
  },
  "paths": {
    "/v1/users/": {
      "post": {
        "tags": ["users"],
        "summary": "Create User",
        "description": "Create a new user in usr.users. Requires full_name, email, auth_user_id, and hashed_password.",
        "operationId": "create_user_v1_users__post",
        "security": [{ "HTTPBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["users"],
        "summary": "List Users",
        "description": "List users from usr.users with pagination.",
        "operationId": "list_users_v1_users__get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserPublic" }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/users/{user_id}": {
      "get": {
        "tags": ["users"],
        "summary": "Get User",
        "description": "Get a user by ID from usr.users.",
        "operationId": "get_user_v1_users__user_id__get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "User Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserPublic" }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Update User",
        "description": "Update a user in usr.users.",
        "operationId": "update_user_v1_users__user_id__put",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "User Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserPublic" }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Soft Delete User",
        "description": "Soft delete a user (sets deleted_at) in usr.users.",
        "operationId": "delete_user_v1_users__user_id__delete",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "User Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "User soft deleted"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserCreate": {
        "type": "object",
        "required": ["full_name", "email", "auth_user_id", "hashed_password"],
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 64,
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "maxLength": 320,
            "title": "Email"
          },
          "auth_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Auth User Id"
          },
          "hashed_password": {
            "type": "string",
            "maxLength": 512,
            "title": "Hashed Password"
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "title": "Phone"
          },
          "avatar_url": {
            "type": "string",
            "maxLength": 512,
            "title": "Avatar URL"
          },
          "permissions": {
            "type": "object",
            "title": "Permissions"
          },
          "is_active": {
            "type": "boolean",
            "default": true,
            "title": "Is Active"
          },
          "is_superuser": {
            "type": "boolean",
            "default": false,
            "title": "Is Superuser"
          }
        },
        "title": "UserCreate",
        "description": "Schema for creating users in usr.users."
      },
      "UserPublic": {
        "type": "object",
        "required": [
          "id",
          "full_name",
          "email",
          "is_active",
          "is_superuser",
          "created_at"
        ],
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "full_name": { "type": "string", "title": "Full Name" },
          "email": { "type": "string", "title": "Email" },
          "phone": { "type": "string", "title": "Phone" },
          "avatar_url": { "type": "string", "title": "Avatar URL" },
          "auth_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Auth User Id"
          },
          "permissions": { "type": "object", "title": "Permissions" },
          "is_active": { "type": "boolean", "title": "Is Active" },
          "is_superuser": { "type": "boolean", "title": "Is Superuser" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Deleted At"
          }
        },
        "title": "UserPublic",
        "description": "Public user schema for API responses from usr.users."
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 64,
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "maxLength": 320,
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "title": "Phone"
          },
          "avatar_url": {
            "type": "string",
            "maxLength": 512,
            "title": "Avatar URL"
          },
          "permissions": {
            "type": "object",
            "title": "Permissions"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser"
          }
        },
        "title": "UserUpdate",
        "description": "Schema for updating users in usr.users."
      },
...
      "HTTPValidationError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" }
          }
        },
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "type": "object",
        "required": ["loc", "msg", "type"],
        "properties": {
          "loc": {
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "msg": { "type": "string" },
          "type": { "type": "string" }
        },
        "title": "ValidationError"
      }
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}